+++ tests/Integration/DSC_DismFeature.Integration.Tests.ps1
@@ -1,0 +1,138 @@
+#region HEADER
+$script:dscModuleName = 'ComputerManagementDsc'
+$script:dscResourceName = 'DSC_DismFeature'
+
+try
+{
+    Import-Module -Name DscResource.Test -Force -ErrorAction 'Stop'
+}
+catch [System.IO.FileNotFoundException]
+{
+    throw 'DscResource.Test module dependency not found. Please run ".\build.ps1 -Tasks build" first.'
+}
+
+$script:testEnvironment = Initialize-TestEnvironment `
+    -DSCModuleName $script:dscModuleName `
+    -DSCResourceName $script:dscResourceName `
+    -ResourceType 'Mof' `
+    -TestType 'Integration'
+
+Import-Module -Name (Join-Path -Path $PSScriptRoot -ChildPath '..\TestHelpers\CommonTestHelper.psm1')
+
+try
+{
+    $configFile = Join-Path -Path $PSScriptRoot -ChildPath "$($script:dscResourceName).Config.ps1"
+    . $configFile
+
+    Describe "$($script:dscResourceName)_Integration" {
+        $featureName = 'IIS-HttpLogging'
+
+        $configData = @{
+            AllNodes = @(
+                @{
+                    NodeName                = 'localhost'
+                    Ensure                  = 'Present'
+                    EnableAllParentFeatures = $true
+                    Name                    = $featureName
+                }
+            )
+        }
+
+        It 'Should compile and apply the MOF without throwing' {
+            {
+                & "$($script:dscResourceName)_Config" `
+                    -OutputPath $TestDrive `
+                    -ConfigurationData $configData
+
+                $startDscConfigurationParameters = @{
+                    Path         = $TestDrive
+                    ComputerName = 'localhost'
+                    Wait         = $true
+                    Verbose      = $true
+                    Force        = $true
+                    ErrorAction  = 'Stop'
+                }
+
+                Start-DscConfiguration @startDscConfigurationParameters
+            } | Should -Not -Throw
+        }
+
+        It 'Should be able to call Get-DscConfiguration without throwing' {
+            {
+                Get-DscConfiguration -Verbose -ErrorAction Stop
+            } | Should -Not -Throw
+        }
+
+        It 'Should have set the resource and all the parameters should match' {
+            $current = Get-DscConfiguration | Where-Object -FilterScript {
+                $_.ConfigurationName -eq "$($script:dscResourceName)_Config"
+            }
+
+            $current.Ensure | Should -Be 'Present'
+            $current.Name | Should -Be $featureName
+            $current.SuppressRestart | Should -BeFalse
+        }
+
+        It 'Should return True when Test-DscConfiguration is run' {
+            Test-DscConfiguration -Verbose | Should -Be 'True'
+        }
+    }
+
+    Describe "$($script:dscResourceName)_Integration" {
+        $featureName = 'IIS-HttpLogging'
+
+        $configData = @{
+            AllNodes = @(
+                @{
+                    NodeName                = 'localhost'
+                    Ensure                  = 'Absent'
+                    EnableAllParentFeatures = $false
+                    Name                    = $featureName
+                }
+            )
+        }
+
+        It 'Should compile and apply the MOF without throwing' {
+            {
+                & "$($script:dscResourceName)_Config" `
+                    -OutputPath $TestDrive `
+                    -ConfigurationData $configData
+
+                $startDscConfigurationParameters = @{
+                    Path         = $TestDrive
+                    ComputerName = 'localhost'
+                    Wait         = $true
+                    Verbose      = $true
+                    Force        = $true
+                    ErrorAction  = 'Stop'
+                }
+
+                Start-DscConfiguration @startDscConfigurationParameters
+            } | Should -Not -Throw
+        }
+
+        It 'Should be able to call Get-DscConfiguration without throwing' {
+            {
+                Get-DscConfiguration -Verbose -ErrorAction Stop
+            } | Should -Not -Throw
+        }
+
+        It 'Should have set the resource and all the parameters should match' {
+            $current = Get-DscConfiguration | Where-Object -FilterScript {
+                $_.ConfigurationName -eq "$($script:dscResourceName)_Config"
+            }
+
+            $current.Ensure | Should -Be 'Absent'
+            $current.Name | Should -Be $featureName
+            $current.SuppressRestart | Should -BeFalse
+        }
+
+        It 'Should return True when Test-DscConfiguration is run' {
+            Test-DscConfiguration -Verbose | Should -Be 'True'
+        }
+    }
+}
+finally
+{
+    Restore-TestEnvironment -TestEnvironment $script:testEnvironment
+}